# Removed `StringBuilder` allocation from `Base64Encoding`