--- a/src/core/Akka/Util/Base64Encoding.cs
+++ b/src/core/Akka/Util/Base64Encoding.cs
@@ -5,6 +5,7 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
+using System;
 using System.Text;
 
 namespace Akka.Util
@@ -24,7 +25,30 @@ namespace Akka.Util
         /// </summary>
         /// <param name="value">TBD</param>
         /// <returns>TBD</returns>
-        public static string Base64Encode(this long value) => Base64Encode(value, new StringBuilder()).ToString();
+        public static string Base64Encode(this long value)
+        {
+            return Base64Encode(value, "");
+        }
+
+        public static string Base64Encode(this long value, string prefix)
+        {
+            Span<char> sb = stackalloc char[11 + prefix?.Length ?? 0];
+            var spanIndex = 0;
+            if (!string.IsNullOrEmpty(prefix) && prefix.Length > 0)
+            {
+                prefix.AsSpan().CopyTo(sb);
+                spanIndex = prefix.Length;
+            }
+
+            var next = value;
+            do
+            {
+                var index = (int)(next & 63);
+                sb[spanIndex++] = Base64Chars[index];
+                next = next >> 6;
+            } while (next != 0);
+            return sb.Slice(0, spanIndex).ToString();
+        }
 
         /// <summary>
         /// TBD
@@ -32,6 +56,7 @@ namespace Akka.Util
         /// <param name="value">TBD</param>
         /// <param name="sb">TBD</param>
         /// <returns>TBD</returns>
+        [Obsolete("Do not use. Pass in prefix as a string instead.")]
         public static StringBuilder Base64Encode(this long value, StringBuilder sb)
         {
             var next = value;